Separated followers controller
