+++ src/routes/user.routes.js
@@ -1,13 +1,13 @@
 import { Router } from "express";
 import { verifyJwtToken } from "../middlewares/auth.middleware.js";
 import {
-  followUnfollowUser,
   getUserProfile,
   loginUser,
   logoutUser,
   registerUser,
 } from "../controllers/user.controller.js";
 import { upload } from "../middlewares/multer.middleware.js";
+import { followUnfollowUser, getFollowersList } from "../controllers/follow.controller.js";
 
 const router = new Router();
 
@@ -20,5 +20,6 @@ router.route("/login").post(loginUser);
 router.route("/logout").post(verifyJwtToken, logoutUser);
 router.route("/:userId").get(verifyJwtToken, getUserProfile);
 router.route("/follow/:userId").post(verifyJwtToken, followUnfollowUser);
+router.route("/followers/:userId").get(verifyJwtToken, getFollowersList);
 
 export default router;
